Added full Test functionality
The entirety of the project is complete for a test GUI. Buffered writers
and readers were replaced for Object output and input streams as the
main method of communication between the server and client is through
Request objects.
Closing clients or servers first requires the instances to close
sockets, readers, and writers first.
This commit is done before actual testing.